Класс для работы с группами
